requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
Brotli==1.0.9
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses==0.8
datasets==2.14.7
dill==0.3.7
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
openai==1.3.6
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pip==23.3.1
protobuf==3.20.3
pyarrow==11.0.0
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sacremoses==0.0.53
safetensors==0.3.3
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sympy==1.12
tokenizers==0.13.3
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.42.0
win-inet-pton==1.1.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
